Rename filesystem plugin to mongodb-docker
If you are still using the filesystem plugin you will see the following
error:

Error: could not list installations: could not load storage plugin: unsupported internal storage plugin specified storage.porter.filesystem
