--- pkg/storage/plugins/filesystem/doc.go
@@ -1,3 +1,0 @@
-// Package filesystem implements the crudstore plugin interface, storing data
-// on the local filesystem using a mini mongo daemon.
-package filesystem
