+++ pkg/storage/plugins/mongodb_docker/doc.go
@@ -1,0 +1,4 @@
+// Package mongodb_docker implements the plugins.StorageProtocol interface, storing data
+// using an instance of mongodb running in a container, with the data stored
+// in a docker volume.
+package mongodb_docker
