+++ pkg/storage/plugins/mongodb_docker/plugin.go
@@ -1,4 +1,4 @@
-package filesystem
+package mongodb_docker
 
 import (
 	"get.porter.sh/porter/pkg/context"
@@ -7,16 +7,16 @@ import (
 	"github.com/pkg/errors"
 )
 
-// PluginKey is the identifier of the internal filesystem plugin.
-const PluginKey = plugins.PluginInterface + ".porter.filesystem"
+// PluginKey is the identifier of the internal mongodb run in docker plugin.
+const PluginKey = plugins.PluginInterface + ".porter.mongodb-docker"
 
-// PluginConfig supported by the filesystem plugin as defined in porter.yaml
+// PluginConfig supported by the mongodb-docker plugin as defined in porter.yaml
 type PluginConfig struct {
 	Port     string `mapstructure:"port"`
 	Database string `mapstructure:"database"`
 }
 
-// NewPlugin creates an instance of the storage.porter.filesystem plugin
+// NewPlugin creates an instance of the storage.porter.mongodb-docker plugin
 func NewPlugin(cxt *context.Context, pluginConfig interface{}) (plugins.StoragePlugin, error) {
 	cfg := PluginConfig{
 		Port:     "27018",
